[Feat -> Dev ] 공연 예외 처리 추가

+++ server/src/main/java/codestates/frogroup/indiego/domain/show/service/ShowService.java
@@ -68,9 +68,12 @@ public class ShowService {
     }
 
     @Transactional
-    public void deleteShow(Long id) {
+    public void deleteShow(Long id, Long memberId) {
 
         Show findShow = findVerifiedShow(id);
+        if(memberId != findShow.getMember().getId()){
+            throw new BusinessLogicException(ExceptionCode.MEMBER_NO_PERMISSION);
+        }
         showRepository.delete(findShow);
 
     }
